Added "Go to (user)" option in more places (posts, comments/mentions/repllies)
Added "Go to (user)" option in more places (posts, comments/mentions/repllies). While this is redundant based on the current UI, it allows for a future setting to turn off inline links for users and communities to stop accidental clickthroughs when users are trying to open comments on a post.