Fix issue of an empty screen in inbox
The issue was caused by the activity checking for a completed request in
the PostViewModel. However, since this activity was triggered by the
inbox, that request was never made.

The only data it needed was whether the user was a moderator, so I pass
that information in with the deep link. In all cases we had enough
information in the navigating activity to determine this, with the
exception of the inbox, but we don't mark moderators in the inbox (at
least lemmy-ui doesn't).

This is a little bit of a hack, but it's a relatively clean hack.