Auto-generate Lemmy-API types. (#657)
* Adding a copy generated types from lemmy-js-client script.

* Initial types commit.

* Changing types to immutable.

* Refactoring to use auto-generated types from ts-rs.

- Also added ApiState to represent 4 API states:
  - Empty, Loading, Fail, and Success
- Still needs testing

* Some more additions.

* Fixing login.

* Resolving a few more issues.

* Using alternate suggestedTitle.

* Compiling, but not fixed.

* Fixing lints.

* Most things working except for subscribe.

* Fixing follow.

* Cleaning up progress indicators.

* Removing unecessary padding.

* Fixing tests.

* Fix crash when marking comment as read (#711)

* Adding word breaks to regex.

* Fix issue of an empty screen in inbox (#715)

The issue was caused by the activity checking for a completed request in
the PostViewModel. However, since this activity was triggered by the
inbox, that request was never made.

The only data it needed was whether the user was a moderator, so I pass
that information in with the deep link. In all cases we had enough
information in the navigating activity to determine this, with the
exception of the inbox, but we don't mark moderators in the inbox (at
least lemmy-ui doesn't).

This is a little bit of a hack, but it's a relatively clean hack.

* Fix invalid navigation route.

* Fix wrong updates on comment edits.

---------

Co-authored-by: Jameson Little <t.jameson.little@gmail.com>