Redesign post listings
To keep the paddings of the composables that make up PostListing
consistent, post activity may look slightly different as well.

Also, moved the clickable in PostListingCard from Column to a new parent
so that the hover effect would affect the entire card.

A goal for this redesign was to reduce clutter by decreasing focus on
post metadata. As a compromise, the user's avatar is no longer shown on
the post.